+++ services/web/server/tests/unit/with_dbs/03/users/test_users_repository.py
@@ -1,0 +1,70 @@
+# pylint: disable=redefined-outer-name
+# pylint: disable=unused-argument
+# pylint: disable=unused-variable
+# pylint: disable=too-many-arguments
+
+from typing import Any
+
+import sqlalchemy as sa
+from aiohttp import web
+from models_library.products import ProductName
+from simcore_postgres_database.models.users_details import (
+    users_pre_registration_details,
+)
+from simcore_service_webserver.db.plugin import get_asyncpg_engine
+from simcore_service_webserver.users._users_repository import (
+    create_user_pre_registration,
+)
+
+
+async def test_create_user_pre_registration(
+    app: web.Application,
+    product_name: ProductName,
+    product_owner_user: dict[str, Any],
+):
+    # Arrange
+    asyncpg_engine = get_asyncpg_engine(app)
+
+    test_email = "test.user@example.com"
+    created_by_user_id = product_owner_user["id"]
+    institution = "Test Institution"
+    pre_registration_details: dict[str, Any] = {
+        "institution": institution,
+        "pre_first_name": "Test",
+        "pre_last_name": "User",
+    }
+
+    # Act
+    pre_registration_id = await create_user_pre_registration(
+        asyncpg_engine,
+        email=test_email,
+        created_by=created_by_user_id,
+        product_name=product_name,
+        **pre_registration_details,
+    )
+
+    # Assert
+    async with asyncpg_engine.connect() as conn:
+        # Query to check if the record was inserted
+        result = await conn.execute(
+            sa.select(users_pre_registration_details).where(
+                (users_pre_registration_details.c.pre_email == test_email)
+                & (users_pre_registration_details.c.product_name == product_name)
+            )
+        )
+        record = result.first()
+
+        # Clean up - delete the test record
+        await conn.execute(
+            sa.delete(users_pre_registration_details).where(
+                users_pre_registration_details.c.id == pre_registration_id
+            )
+        )
+        await conn.commit()
+
+    # Verify the record was created with correct values
+    assert record is not None
+    assert record.pre_email == test_email
+    assert record.created_by == created_by_user_id
+    assert record.product_name == product_name
+    assert record.institution == institution
